requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.0.1
av==15.1.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.9
dill==0.3.8
docstring_parser==0.17.0
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.118.0
ffmpy==0.6.1
filelock==3.19.1
fire==0.7.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.3.0
gradio==5.45.0
gradio_client==1.13.0
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.1.10
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
-e git+https://github.com/hiyouga/LLaMA-Factory.git@d5bb4e6394409394c4a892571a5bc94fadd874c7#egg=llamafactory
llvmlite==0.45.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
modelscope==1.30.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.7.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
orjson==3.11.3
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
propcache==0.4.0
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qwen-vl-utils==0.0.8
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rouge-chinese==1.0.3
ruff==0.13.3
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
transformers-stream-generator==0.0.5
triton==3.4.0
trl==0.9.6
typer==0.19.2
typing_extensions==4.15.0
tyro==0.8.14
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
